Refactor and move common code.
